+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add colon to allowed characters of leaf interface selector names
 - Normalize access SPAN filter `protocol` and `port` values
+- Add support for `auto_generate_access_leaf_switch_interface_profiles` and `auto_generate_access_spine_switch_interface_profiles` flags
 
 ## 0.3.1